license-plate-detection/requirements.txt
ultralytics
opencv-python
numpy
Pillow
torch
torchvision
matplotlib